+++ src/app/panel_state.rs
@@ -490,6 +490,9 @@ pub trait PanelState {
                     CmdResult::Keep
                 }
             }
+            Internal::escape => {
+                CmdResult::HandleInApp(Internal::escape)
+            }
             Internal::panel_left | Internal::panel_left_no_open => {
                 CmdResult::HandleInApp(Internal::panel_left_no_open)
             }
@@ -749,11 +752,11 @@ pub trait PanelState {
                 }
             }
             Command::VerbTrigger {
-                index,
+                verb_id,
                 input_invocation,
             } => self.execute_verb(
                 w,
-                &con.verb_store.verbs[*index],
+                con.verb_store.verb(*verb_id),
                 input_invocation.as_ref(),
                 TriggerType::Other,
                 app_state,