+++ src/command/panel_input.rs
@@ -78,14 +78,22 @@ impl PanelInput {
     pub fn on_event(
         &mut self,
         w: &mut W,
-        event: TimedEvent,
+        timed_event: TimedEvent,
         con: &AppContext,
         sel_info: SelInfo<'_>,
         app_state: &AppState,
         mode: Mode,
         panel_state_type: PanelStateType,
     ) -> Result<Command, ProgramError> {
-        let cmd = self.get_command(event, con, sel_info, app_state, mode, panel_state_type);
+        let cmd = match timed_event.event {
+            Event::Mouse(MouseEvent { kind, column, row, modifiers: KeyModifiers::NONE }) => {
+                self.on_mouse(timed_event, kind, column, row)
+            }
+            Event::Key(key) => {
+                self.on_key(timed_event, key, con, sel_info, app_state, mode, panel_state_type)
+            }
+            _ => Command::None,
+        };
         self.input_field.display_on(w)?;
         Ok(cmd)
     }
@@ -181,13 +189,15 @@ impl PanelInput {
 
     /// escape (bound to the 'esc' key)
     ///
-    /// Can't be done as a standard internal for the moment because of
-    /// the necessary position before the 'tab' if/else handling
-    fn escape(
+    /// This function is better called from the on_key method of
+    /// panel input, when a key triggers it, because then it
+    /// can also properly deal with completion sequence.
+    /// When ':escape' is called from a verb's cmd sequence, then
+    /// it's not called on on_key but by the app.
+    pub fn escape(
         &mut self,
         con: &AppContext,
         mode: Mode,
-        parts: CommandParts,
     ) -> Command {
         self.tab_cycle_count = 0;
         if let Some(raw) = self.input_before_cycle.take() {
@@ -203,6 +213,8 @@ impl PanelInput {
             }
         } else {
             // general back command
+            let raw = self.input_field.get_content();
+            let parts = CommandParts::from(raw.clone());
             self.input_field.clear();
             let internal = Internal::back;
             Command::Internal {
@@ -268,8 +280,8 @@ impl PanelInput {
         con: &'c AppContext,
         sel_info: SelInfo<'_>,
         panel_state_type: PanelStateType,
-    ) -> Option<(usize, &'c Verb)> {
-        for (index, verb) in con.verb_store.verbs.iter().enumerate() {
+    ) -> Option<&'c Verb> {
+        for verb in con.verb_store.verbs().iter() {
             // note that there can be several verbs with the same key and
             // not all of them can apply
             if !verb.keys.contains(&key) {
@@ -288,135 +300,149 @@ impl PanelInput {
                 }
             }
             debug!("verb for key: {}", &verb.execution);
-            return Some((index, verb));
+            return Some(verb);
         }
         None
     }
 
-    /// Consume the event to
-    /// - maybe change the input
-    /// - build a command
-    fn get_command(
+    /// Consume the event, maybe change the input, return a command
+    fn on_mouse(
         &mut self,
         timed_event: TimedEvent,
+        kind: MouseEventKind,
+        column: u16,
+        row: u16,
+    ) -> Command {
+        if self.input_field.apply_timed_event(timed_event) {
+            Command::empty()
+        } else {
+            match kind {
+                MouseEventKind::Up(MouseButton::Left) => {
+                    if timed_event.double_click {
+                        Command::DoubleClick(column, row)
+                    } else {
+                        Command::Click(column, row)
+                    }
+                }
+                MouseEventKind::ScrollDown => {
+                    Command::Internal {
+                        internal: Internal::line_down,
+                        input_invocation: None,
+                    }
+                }
+                MouseEventKind::ScrollUp => {
+                    Command::Internal {
+                        internal: Internal::line_up,
+                        input_invocation: None,
+                    }
+                }
+                _ => Command::None,
+            }
+        }
+    }
+
+    /// Consume the event, maybe change the input, return a command
+    #[allow(clippy::too_many_arguments)]
+    fn on_key(
+        &mut self,
+        timed_event: TimedEvent,
+        key: KeyEvent,
         con: &AppContext,
         sel_info: SelInfo<'_>,
         app_state: &AppState,
         mode: Mode,
         panel_state_type: PanelStateType,
     ) -> Command {
-        match timed_event.event {
-            Event::Mouse(MouseEvent { kind, column, row, modifiers: KeyModifiers::NONE }) => {
-                if self.input_field.apply_timed_event(timed_event) {
-                    Command::empty()
-                } else {
-                    match kind {
-                        MouseEventKind::Up(MouseButton::Left) => {
-                            if timed_event.double_click {
-                                Command::DoubleClick(column, row)
-                            } else {
-                                Command::Click(column, row)
-                            }
-                        }
-                        MouseEventKind::ScrollDown => {
-                            Command::Internal {
-                                internal: Internal::line_down,
-                                input_invocation: None,
-                            }
-                        }
-                        MouseEventKind::ScrollUp => {
-                            Command::Internal {
-                                internal: Internal::line_up,
-                                input_invocation: None,
-                            }
-                        }
-                        _ => Command::None,
-                    }
-                }
-            }
-            Event::Key(key) => {
-                // value of raw and parts before any key related change
-                let raw = self.input_field.get_content();
-                let mut parts = CommandParts::from(raw.clone());
+        // value of raw and parts before any key related change
+        let raw = self.input_field.get_content();
+        let parts = CommandParts::from(raw.clone());
 
-                // We first handle the cases that MUST absolutely
-                // not be overridden by configuration.
-                // Those handlings must also be done in order: escaping
-                // a cycling must be done before the 'tab' handling (and
-                // its else which resets the cycle)
+        let verb = if keys::is_key_allowed_for_verb(key, mode, raw.is_empty()) {
+            self.find_key_verb(
+                key,
+                con,
+                sel_info,
+                panel_state_type,
+            )
+        } else {
+            None
+        };
 
-                if key == key!(esc) {
-                    return self.escape(con, mode, parts);
-                }
+        // WARNINGS:
+        // - beware the execution order below: we must execute
+        // escape before the else clause of next_match, and we must
+        // be sure this else clause (which ends cycling) is always
+        // executed when neither next_match or escape is triggered
+        // - some behaviors can't really be handled as normally
+        // triggered internals because of the interactions with
+        // the input
 
-                // tab completion
-                if key == key!(tab) {
-                    if parts.verb_invocation.is_some() {
-                        return self.auto_complete_verb(con, sel_info, raw, parts);
-                    }
-                    // if no verb is being edited, other bindings may apply (eg :next_match)
-                } else {
-                    self.tab_cycle_count = 0;
-                    self.input_before_cycle = None;
-                }
-
-                if key == key!(enter) && parts.has_not_empty_verb_invocation() {
-                    return Command::from_parts(parts, true);
-                }
-
-                if (key == key!('?') || key == key!(shift-'?'))
-                    && (raw.is_empty() || parts.verb_invocation.is_some()) {
-                    // a '?' opens the help when it's the first char
-                    // or when it's part of the verb invocation
-                    return Command::Internal {
-                        internal: Internal::help,
-                        input_invocation: parts.verb_invocation,
-                    };
-                }
-
-                // we now check if the key is the trigger key of one of the verbs
-                if keys::is_key_allowed_for_verb(key, mode, raw.is_empty()) {
-                    if let Some((index, verb)) = self.find_key_verb(
-                        key,
-                        con,
-                        sel_info,
-                        panel_state_type,
-                    ) {
-                        if self.handle_input_related_verb(verb, con) {
-                            return Command::from_raw(self.input_field.get_content(), false);
-                        }
-                        if mode != Mode::Input && verb.is_internal(Internal::mode_input) {
-                            self.enter_input_mode_with_key(key, &parts);
-                        }
-                        if verb.auto_exec {
-                            return Command::VerbTrigger {
-                                index,
-                                input_invocation: parts.verb_invocation,
-                            };
-                        }
-                        if let Some(invocation_parser) = &verb.invocation_parser {
-                            let exec_builder = ExecutionStringBuilder::without_invocation(
-                                sel_info,
-                                app_state,
-                            );
-                            let verb_invocation = exec_builder.invocation_with_default(
-                                &invocation_parser.invocation_pattern
-                            );
-                            parts.verb_invocation = Some(verb_invocation);
-                            self.set_content(&parts.to_string());
-                            return Command::VerbEdit(parts.verb_invocation.unwrap());
-                        }
-                    }
-                }
-
-                // input field management
-                if mode == Mode::Input && self.input_field.apply_timed_event(timed_event) {
-                    return Command::from_raw(self.input_field.get_content(), false);
-                }
-                Command::None
-            }
-            _ => Command::None,
+        // usually 'esc' key
+        if Verb::is_some_internal(verb, Internal::escape) {
+            return self.escape(con, mode);
         }
+
+        // 'tab' completion of a verb or one of its arguments
+        if Verb::is_some_internal(verb, Internal::next_match) {
+            if parts.verb_invocation.is_some() {
+                return self.auto_complete_verb(con, sel_info, raw, parts);
+            }
+            // if no verb is being edited, the state may handle this internal
+            // in a specific way
+        } else {
+            self.tab_cycle_count = 0;
+            self.input_before_cycle = None;
+        }
+
+        // 'enter': trigger the verb if any on the input. If none, then may be
+        // used as trigger of another verb
+        if key == key!(enter) && parts.has_not_empty_verb_invocation() {
+            return Command::from_parts(parts, true);
+        }
+
+        // a '?' opens the help when it's the first char or when it's part
+        // of the verb invocation. It may be used as a verb name in other cases
+        if (key == key!('?') || key == key!(shift-'?'))
+            && (raw.is_empty() || parts.verb_invocation.is_some()) {
+            return Command::Internal {
+                internal: Internal::help,
+                input_invocation: parts.verb_invocation,
+            };
+        }
+
+        if let Some(verb) = verb {
+            if self.handle_input_related_verb(verb, con) {
+                return Command::from_raw(self.input_field.get_content(), false);
+            }
+            if mode != Mode::Input && verb.is_internal(Internal::mode_input) {
+                self.enter_input_mode_with_key(key, &parts);
+            }
+            if verb.auto_exec {
+                return Command::VerbTrigger {
+                    verb_id: verb.id,
+                    input_invocation: parts.verb_invocation,
+                };
+            }
+            if let Some(invocation_parser) = &verb.invocation_parser {
+                let exec_builder = ExecutionStringBuilder::without_invocation(
+                    sel_info,
+                    app_state,
+                );
+                let verb_invocation = exec_builder.invocation_with_default(
+                    &invocation_parser.invocation_pattern
+                );
+                let mut parts = parts;
+                parts.verb_invocation = Some(verb_invocation);
+                self.set_content(&parts.to_string());
+                return Command::VerbEdit(parts.verb_invocation.unwrap());
+            }
+        }
+
+        // input field management
+        if mode == Mode::Input && self.input_field.apply_timed_event(timed_event) {
+            return Command::from_raw(self.input_field.get_content(), false);
+        }
+        Command::None
     }
 }