+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### next
+- allow rebinding of the 'tab' and 'esc' keys with the `:next_match` and `:escape` internals - Fix #740
 - fix fuzzy patterns not case insensitive on some characters - Fix #746
 - when given a path to a file at launch, broot now selects it in the tree and opens it in preview - Fix #729